Add files via upload
The automated service and maintenance team of 301 Publishing House has developed a four-legged walking robot mechanism based on a series of belt drives and clutches. The robot mimics the motion of quadrupeds in nature and is capable of maneuvering complex terrains and carrying light loads, allowing it to be used for searching and maintenance activities in small spaces with height differences across the shop floor. Upon learning about the team’s expertise in walking robots, the shipping and distribution department of 301 Publishing House requested a modified mechanism to transport and lift packaged boxes of books for stacking on shipping pallets.